Bug related to improper handling of forwarding fixed.
